Added tags @sigma and @imperial
Please read the comments in the document.

The changes from the new correction_factor function (affecting sigma, sigmat, and sigmai) are now being traced to all other functions/variables affected by it. The "@sigma" tag will appear in comments near these occurrences. 

"@imperial" tags are being used to trace variables and make sure a conversion to metric is as smooth as possible.